+++ opentelemetry-api/tests/propagators/test_global_httptextformat.py
@@ -16,8 +16,8 @@ import unittest
 
 from opentelemetry import baggage, trace
 from opentelemetry.propagate import extract, inject
+from opentelemetry.propagators.textmap import DictGetter
 from opentelemetry.trace import get_current_span, set_span_in_context
-from opentelemetry.trace.propagation.textmap import DictGetter
 
 carrier_getter = DictGetter()
 
@@ -48,7 +48,7 @@ class TestDefaultGlobalPropagator(unittest.TestCase):
         self.assertEqual(span_context.trace_id, self.TRACE_ID)
         self.assertEqual(span_context.span_id, self.SPAN_ID)
 
-        span = trace.DefaultSpan(span_context)
+        span = trace.NonRecordingSpan(span_context)
         ctx = baggage.set_baggage("key3", "val3")
         ctx = baggage.set_baggage("key4", "val4", context=ctx)
         ctx = set_span_in_context(span, context=ctx)